Add book loan management and refactor existing classes
New Features:
- Added Loan model and LoanStatus enum to represent book loans.
- Implemented LoanRepository and LoanRepositoryInterface for database operations on loans.
- Introduced LoanService to handle borrowing and returning books.

Changes to Existing Classes:
- Refactored LibraryService – moved borrowing/return logic to LoanService.
- Updated UserService – integrated with LoanService to check active loans.
- Changed userId type in User class from Integer to Long.
- Adjusted UserRepository and BookRepository (ID type changes, SQL query fixes).

Project Configuration:
- Added .idea/data_source_mapping.xml and .idea/sqldialects.xml for data source and SQL dialect mappings.

+++ src/main/java/org/library/model/User.java
@@ -10,7 +10,7 @@ import java.util.List;
 @ToString(exclude = {"borrowedBooks"})
 @RequiredArgsConstructor
 public class User {
-    private Integer userId;
+    private Long userId;
     @NonNull
     private String fullName;
     @NonNull
@@ -25,7 +25,7 @@ public class User {
         this.password = password;
     }
 
-    public User(Integer userId, @NonNull String fullName, @NonNull String email, String password) {
+    public User(Long userId, @NonNull String fullName, @NonNull String email, String password) {
         this.userId = userId;
         this.fullName = fullName;
         this.email = email;
